% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_fit.R
\name{stat_fit}
\alias{stat_fit}
\title{Title}
\usage{
stat_fit(
  mapping = NULL,
  data = NULL,
  geom = "point",
  position = "identity",
  ...,
  method = NULL,
  formula = NULL,
  se = TRUE,
  n = 80,
  span = 0.75,
  fullrange = FALSE,
  level = 0.95,
  method.args = list(),
  na.rm = FALSE,
  orientation = NA,
  show.legend = NA,
  inherit.aes = TRUE
)
}
\arguments{
\item{mapping}{}

\item{data}{}

\item{geom}{}

\item{position}{}

\item{...}{}

\item{method}{}

\item{formula}{}

\item{se}{}

\item{n}{}

\item{span}{}

\item{fullrange}{}

\item{level}{}

\item{method.args}{}

\item{na.rm}{}

\item{orientation}{}

\item{show.legend}{}

\item{inherit.aes}{}
}
\description{
Title
}
